unsigned log => unsigned int (32-bit) for 64-bit Linux

+++ bigdecimal.h
@@ -64,7 +64,8 @@ extern "C" {
 /* VP_HANDLE,VP_DIGIT,VP_UINT must be unsigned. */
 /***/
 #define VP_HANDLE  unsigned long long  /* This is actually an address for structure Real. */
-#define VP_DIGIT   unsigned long       /* Fraction part array (can be 64-bit:unsigned long long)   */
+/* unsigned log is 32-bit on Windows,but it is 64-bit on 64-bit Linux */
+#define VP_DIGIT   unsigned int        /* Fraction part array (can be 64-bit:unsigned long long)   */
 #define VP_UINT    size_t              /* Common unsigned part   */
 
 /* VP-exception handle called from VP-routines if needed(error case). */
